refactor(storage): split `WellKnownParameter`
