+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -108,6 +108,7 @@ google_cloud_cpp_storage_hdrs = [
     "internal/tracing_connection.h",
     "internal/tuple_filter.h",
     "internal/unified_rest_credentials.h",
+    "internal/well_known_parameters_impl.h",
     "lifecycle_rule.h",
     "list_buckets_reader.h",
     "list_hmac_keys_reader.h",
